fix: move @braw/extended-intl to peer dependencies
BREAKING CHANGE: @braw/extended-intl is no longer listed as a
dependency, but rather a peer dependency. That's because extended-intl
comes with locale data imports that can be used by consumer of this
library, therefore it makes sense to install it manually instead of
relying on hoisting. To migrate, simply list `@braw/extended-intl` as
your package dependency.