feat: add default message ordering
Default message is now fetched from either a descriptor or the loaded
default locale's messages, with the defined order through config.

BREAKING CHANGE: Since default message is now fetched as defined in
options, and the default for that option is fetching messages from
descriptor only, if you relied on the behaviour of using the default
message from the loaded default locale's messages, you may want to
change `defaultMessageOrder` option to `['locale', 'descriptor']`.